Moving metadata into setup.cfg leaving setup.py as completely templatable
