+++ {{ cookiecutter.project_slug }}/setup.cfg
@@ -1,1 +1,30 @@
-../setup.cfg
+# https://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
+[metadata]
+name = {{ cookiecutter.project_slug }}
+version = 1.0
+description = {{ cookiecutter.description }}
+long_description = file: README.md
+long_description_content_type = text/markdown
+
+author = Hypothesis Engineering Team
+author_email = eng@list.hypothes.is
+maintainer = Hypothesis Engineering Team
+maintainer_email = eng@list.hypothes.is
+url = https://web.hypothes.is/
+
+# From: https://pypi.org/pypi?:action=list_classifiers
+classifiers =
+    Development Status :: 2 - Pre-Alpha
+    Programming Language :: Python :: 3.6
+
+project_urls=
+    Source=https://github.com/hypothesis/h-cookiecutter-pypackage
+license=License :: OSI Approved :: BSD License
+platforms=Operating System :: OS Independent
+
+[options]
+install_requires =
+tests_require=
+    pytest
+    coverage
+
