--- {{ cookiecutter.project_slug }}/setup_extra.py
@@ -1,1 +1,0 @@
-../setup_extra.py
